refactor: viewModel 파일 분리
